# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Gerador de Boletos",
    "description": "API para geração de dados de boletos bancários (Linha digitável e Código de Barras)"
  },
  "paths": {
    "/api/boleto/generate": {
      "post": {
        "summary": "Gera dados de um boleto",
        "description": "Cria linha digitável e código de barras baseado nos parâmetros fornecidos ou gera dados aleatórios.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "banco": {
                    "type": "string",
                    "enum": [
                      "bradesco",
                      "caixa",
                      "itau",
                      "santander"
                    ]
                  },
                  "dataEmissao": {
                    "type": "string",
                    "format": "date"
                  },
                  "dataVencimento": {
                    "type": "string",
                    "format": "date"
                  },
                  "valorDocumento": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "nossoNumero": {
                    "type": "string"
                  },
                  "agencia": {
                    "type": "string"
                  },
                  "codigoCedente": {
                    "type": "string"
                  },
                  "carteira": {
                    "type": "string"
                  },
                  "localPagamento": {
                    "type": "string"
                  },
                  "numeroDocumento": {
                    "type": "string"
                  },
                  "cedente": {
                    "type": "string"
                  },
                  "cedenteCnpj": {
                    "type": "string"
                  },
                  "instrucoesPagamento": {
                    "type": "string"
                  },
                  "identificadorEmissao": {
                    "type": "string"
                  }
                },
                "additionalProperties": false,
                "$schema": "http://json-schema.org/draft-07/schema#"
              },
              "example": {
                "banco": "itau",
                "dataEmissao": "2024-01-01",
                "dataVencimento": "2024-02-01",
                "valorDocumento": 10000,
                "nossoNumero": "123456789",
                "agencia": "1234",
                "codigoCedente": "12345",
                "carteira": "175",
                "localPagamento": "Pagável em qualquer banco até o vencimento",
                "numeroDocumento": "123456",
                "cedente": "Empresa Exemplo LTDA",
                "cedenteCnpj": "12345678000190",
                "instrucoesPagamento": "Não receber após o vencimento",
                "identificadorEmissao": "1"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Boleto gerado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigoBarras": {
                      "type": "string"
                    },
                    "linhaDigitavel": {
                      "type": "string"
                    },
                    "banco": {
                      "type": "string"
                    },
                    "codigoBanco": {
                      "type": "string"
                    },
                    "dataEmissao": {
                      "type": "string"
                    },
                    "dataVencimento": {
                      "type": "string"
                    },
                    "valorDocumento": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "codigoBarras",
                    "linhaDigitavel",
                    "banco",
                    "codigoBanco",
                    "dataEmissao",
                    "dataVencimento",
                    "valorDocumento"
                  ],
                  "example": {
                    "codigoBarras": "34191592000000100001000123456789017500000000000",
                    "linhaDigitavel": "34191.00012 00000.10001 01234.567890 1 75100000000000",
                    "banco": "itau",
                    "codigoBanco": "341",
                    "dataEmissao": "2024-01-01",
                    "dataVencimento": "2024-02-01",
                    "valorDocumento": 10000
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}